only admin users can delete employees

+++ routes/web.php
@@ -13,7 +13,7 @@ Route::middleware(['auth', 'verified'])->group(function () {
     Route::get('employees', [EmployeeController::class, 'index'])->name('employees.index');
     Route::post('employees', [EmployeeController::class, 'store'])->name('employees.store');
     Route::put('employees/{employee}', [EmployeeController::class, 'update'])->name('employees.update');
-    Route::delete('employees/{employee}', [EmployeeController::class, 'destroy'])->name('employees.destroy');
+    Route::delete('employees/{employee}', [EmployeeController::class, 'destroy'])->middleware('can:access-admin-ui')->name('employees.destroy');
     Route::get('calculate', [CalculateController::class, 'index'])->name('calculate.index');
     Route::post('calculate', [CalculateController::class, 'store'])->name('calculate.store');
     Route::get('summary', [SummaryController::class, 'index'])->name('summary.index');
